[ASI-587] Refactor random utils used by both mad dog and seed into service commands to avoid circular dependency with mad dog
