+++ service-commands/src/commands/seed/index.js
@@ -1,0 +1,3 @@
+const SeedSession = require('./SeedSession')
+
+module.exports = SeedSession
